+++ .github/workflows/cd-rinkeby.yml
@@ -32,11 +32,11 @@ jobs:
     - name: Install dependencies
       run: yarn install --frozen-lockfile
     - name: Deploy Factory
-      run: cd packages/govern-core && yarn deploy-factory --network rinkeby
+      run: cd packages/govern-create && yarn deploy-factory --network rinkeby
     - name: Deploy Govern
-      run: cd packages/govern-core && yarn deploy-govern --network rinkeby > github-message
+      run: cd packages/govern-create && yarn deploy-govern --network rinkeby > github-message
     - name: Log
-      run: cat packages/govern-core/github-message
+      run: cat packages/govern-create/github-message
     - name: Shill
       uses: thollander/actions-comment-pull-request@1.0.0
       if: github.event_name == 'pull_request'
