+++ packages/govern-create/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@aragon/govern-create",
+  "version": "1.0.0-beta.0",
+  "description": "Factories and utilities to deploy the Govern system",
+  "main": "index.js",
+  "repository": "https://github.com/aragon/govern",
+  "author": "Aragon Association <legal@aragon.org>",
+  "contributors": [
+    "Jorge Izquierdo <jorge@aragon.one>"
+  ],
+  "files": [
+    "/artifacts",
+    "/contracts"
+  ],
+  "license": "GPL-3.0",
+  "private": false,
+  "dependencies": {
+    "@aragon/govern-core": "^1.0.0",
+    "@nomiclabs/buidler-etherscan": "^2.1.0",
+    "dotenv": "^8.2.0",
+    "node-fetch": "^2.6.1",
+    "unique-names-generator": "^4.3.1",
+    "web3-eth-abi": "^1.3.0"
+  },
+  "devDependencies": {
+    "@nomiclabs/buidler": "^1.4.7",
+    "@nomiclabs/buidler-ethers": "^2.0.0",
+    "@nomiclabs/buidler-waffle": "^2.0.0",
+    "chai": "^4.2.0",
+    "ethereum-waffle": "^3.0.0",
+    "ethers": "^5.0.0"
+  },
+  "scripts": {
+    "clean": "rm -rf artifacts",
+    "coverage": "SOLIDITY_COVERAGE=true buidler coverage --network coverage",
+    "compile": "buidler compile --force",
+    "deploy-factory": "yarn compile && buidler deploy-factory",
+    "deploy-govern": "buidler deploy-govern",
+    "test": "buidler test",
+    "prepublishOnly": "yarn test && yarn compile"
+  }
+}
