+++ packages/govern-subgraph/manifest/data/mainnet-staging.json
@@ -1,6 +1,6 @@
 {
   "network": "mainnet",
   "dataSources": {
-    "GovernFactory": []
+    "GovernBaseFactory": []
   }
 }
