+++ packages/govern-subgraph/manifest/data/mainnet.json
@@ -1,6 +1,6 @@
 {
   "network": "mainnet",
   "dataSources": {
-    "GovernFactory": []
+    "GovernBaseFactory": []
   }
 }
